Fixed some stuff, made a few small changes, removed Java 16 code
Before it was not working because of multiple issues with the Classloader which I have fixed. Its not perfect, but it at least boots into the game which it could not before.
I have also removed some of the Java 16 and newer specific code to work on Java 8 and 11.
This updated version was able to Launch on TL with my RHEL9 system. It even said modded (which for some reason it did not on Vanilla FC despite having similar JSON).
Anyhow I hope this works for you.